U-Boot Firmware Version Comparison
This is a rough idea of how the Firmware Version Comparison feature
will function. In the next installments, I will be adding this function
into the U-Boot process. After coding this up, I may just move to C
Regex to simplify the process, although, it wouldn’t provide detailed
reasons for why the script won’t function. Lastly, this has a small bug
in the for loops that I’ll fix later.