Regex Version Comparison
This script is a replacement for the previous version checking script.
Ideally, the use of a regex syntax will allow for easier readability as
well as limit the potential edge cases that I will encounter. This
script is the basis for what will be added into U-Boot. The only
difference will be analyzing memory locations rather than command line
inputs.